Add keyboard shortcut into context menus
